implement better validation approach and coverage of openapi spec, some dto tweaks

+++ packages/core/src/dtos/request/acceptance.rs
@@ -1,4 +1,5 @@
 use chrono::{DateTime, Utc};
+use sea_orm::entity::prelude::Decimal;
 use uuid::Uuid;
 use voletu_core_macros::request_dto;
 
@@ -10,6 +11,7 @@ pub struct CreateAcceptanceRequest {
   pub document_number: String,
   pub date_accepted: DateTime<Utc>,
   pub arrival_type: ArrivalType,
+  #[validate(length(min = 1))]
   pub source_entity: Option<String>,
   pub truck_waybill_id: Option<Uuid>,
   pub rail_waybill_id: Option<Uuid>,
@@ -21,7 +23,7 @@ pub struct CreateAcceptanceItemRequest {
   pub acceptance_doc_id: Uuid,
   pub product_id: Uuid,
   pub contractor_id: Uuid,
-  pub accepted_amount: f64,
+  pub accepted_amount: Decimal,
 }
 
 #[request_dto]
@@ -34,14 +36,14 @@ pub struct CreateAcceptanceAllocationRequest {
 #[request_dto]
 pub struct AcceptanceAllocationCompositeRequest {
   pub storage_id: Uuid,
-  pub allocated_amount: f64,
+  pub allocated_amount: Decimal,
 }
 
 #[request_dto]
 pub struct AcceptanceItemCompositeRequest {
   pub product_id: Uuid,
   pub contractor_id: Uuid,
-  pub accepted_amount: f64,
+  pub accepted_amount: Decimal,
   #[validate(length(min = 1))]
   pub allocations: Vec<AcceptanceAllocationCompositeRequest>,
 }
