+++ packages/core/src/dtos/response/ledger.rs
@@ -1,3 +1,4 @@
+use sea_orm::entity::prelude::Decimal;
 use uuid::Uuid;
 use voletu_core_macros::response_dto;
 
@@ -6,5 +7,5 @@ pub struct LedgerEntryResponse {
   pub storage_id: Uuid,
   pub product_id: Uuid,
   pub contractor_id: Uuid,
-  pub current_amount: f64,
+  pub current_amount: Decimal,
 }
