+++ packages/core/src/dtos/response/operations.rs
@@ -1,3 +1,4 @@
+use sea_orm::entity::prelude::Decimal;
 use uuid::Uuid;
 use voletu_core_macros::response_dto;
 
@@ -14,7 +15,7 @@ pub struct PhysicalTransferResponse {
   pub product_id: Uuid,
   pub from_storage_id: Uuid,
   pub to_storage_id: Uuid,
-  pub amount_transferred: f64,
+  pub amount_transferred: Decimal,
 }
 
 #[response_dto]
@@ -25,7 +26,7 @@ pub struct OwnershipTransferResponse {
   pub product_id: Uuid,
   pub from_contractor_id: Uuid,
   pub to_contractor_id: Uuid,
-  pub amount_transferred: f64,
+  pub amount_transferred: Decimal,
 }
 
 #[response_dto]
@@ -43,7 +44,7 @@ pub struct BlendingComponentResponse {
   pub blending_doc_id: Uuid,
   pub storage_id: Uuid,
   pub source_product_id: Uuid,
-  pub amount_used: f64,
+  pub amount_used: Decimal,
 }
 
 #[response_dto]
@@ -51,7 +52,7 @@ pub struct BlendingResultResponse {
   pub id: Uuid,
   pub blending_doc_id: Uuid,
   pub storage_id: Uuid,
-  pub produced_amount: f64,
+  pub produced_amount: Decimal,
 }
 
 #[response_dto]
@@ -70,7 +71,7 @@ pub struct InventoryAdjustmentResponse {
   pub product_id: Uuid,
   pub contractor_id: Uuid,
   pub adjustment_type: AdjustmentType,
-  pub amount: f64,
+  pub amount: Decimal,
   pub reason: Option<String>,
 }
